+++ src/Container/Container.php
@@ -18,16 +18,13 @@ use function is_string;
 
 class Container implements ContainerInterface
 {
-    private ?DependencyResolver $dependencyResolver = null;
-
-    /** @var array<class-string|string, list<mixed>> */
-    private array $cachedDependencies = [];
-
     private AliasRegistry $aliasRegistry;
 
     private FactoryManager $factoryManager;
 
     private InstanceRegistry $instanceRegistry;
+
+    private DependencyCacheManager $cacheManager;
 
     /**
      * @param  array<class-string, class-string|callable|object>  $bindings
@@ -40,6 +37,7 @@ class Container implements ContainerInterface
         $this->aliasRegistry = new AliasRegistry();
         $this->factoryManager = new FactoryManager($instancesToExtend);
         $this->instanceRegistry = new InstanceRegistry();
+        $this->cacheManager = new DependencyCacheManager($bindings);
     }
 
     /**
@@ -84,16 +82,12 @@ class Container implements ContainerInterface
     public function resolve(callable $callable): mixed
     {
         $callableKey = $this->callableKey($callable);
-        $callable = Closure::fromCallable($callable);
+        $closure = Closure::fromCallable($callable);
 
-        if (!isset($this->cachedDependencies[$callableKey])) {
-            $this->cachedDependencies[$callableKey] = $this
-                ->getDependencyResolver()
-                ->resolveDependencies($callable);
-        }
+        $dependencies = $this->cacheManager->resolveCallableDependencies($callableKey, $closure);
 
         /** @psalm-suppress MixedMethodCall */
-        return $callable(...$this->cachedDependencies[$callableKey]);
+        return $closure(...$dependencies);
     }
 
     public function factory(Closure $instance): Closure
@@ -208,18 +202,7 @@ class Container implements ContainerInterface
      */
     public function warmUp(array $classNames): void
     {
-        foreach ($classNames as $className) {
-            if (!class_exists($className)) {
-                continue;
-            }
-
-            // Pre-resolve dependencies to populate cache
-            if (!isset($this->cachedDependencies[$className])) {
-                $this->cachedDependencies[$className] = $this
-                    ->getDependencyResolver()
-                    ->resolveDependencies($className);
-            }
-        }
+        $this->cacheManager->warmUp($classNames);
     }
 
     private function createInstance(string $class): ?object
@@ -252,29 +235,7 @@ class Container implements ContainerInterface
      */
     private function instantiateClass(string $class): ?object
     {
-        if (class_exists($class)) {
-            if (!isset($this->cachedDependencies[$class])) {
-                $this->cachedDependencies[$class] = $this
-                    ->getDependencyResolver()
-                    ->resolveDependencies($class);
-            }
-
-            /** @psalm-suppress MixedMethodCall */
-            return new $class(...$this->cachedDependencies[$class]);
-        }
-
-        return null;
-    }
-
-    private function getDependencyResolver(): DependencyResolver
-    {
-        if ($this->dependencyResolver === null) {
-            $this->dependencyResolver = new DependencyResolver(
-                $this->bindings,
-            );
-        }
-
-        return $this->dependencyResolver;
+        return $this->cacheManager->instantiate($class);
     }
 
     /**